test: consolidate chat history paused e2e tests into `chat.spec.ts`
