+++ apps/frontend/tests/e2e/chat.spec.ts
@@ -974,6 +974,151 @@ test.describe("Chat", () => {
 		},
 	);
 
+	testWithMockedLlm(
+		"Activating web search shows chat history paused info message in a chat with history",
+		async ({ page, isMobile }) => {
+			testWithMockedLlm.skip(isMobile === true, "desktop-only menu flow");
+
+			if (process.env.VITE_FEATURE_FLAG_WEB_SEARCH_ALLOWED !== "true") {
+				testWithMockedLlm.skip();
+			}
+
+			await page.goto("/");
+
+			const chatInput = page.getByPlaceholder("Stellen Sie eine Frage");
+			await chatInput.fill("Hallo, wie geht es dir?");
+			await sendAndWaitForLLMResponse(page);
+
+			const answer = page.getByTestId("assistant-message-markdown-container");
+			await expect(answer).not.toBeEmpty();
+
+			const chatOptionsButton = page.getByRole("button", {
+				name: "Weitere Funktionen aktivieren",
+			});
+			await chatOptionsButton.click();
+
+			const webSearchOption = page.getByRole("menuitemcheckbox", {
+				name: "Websuche auswählen",
+			});
+			await webSearchOption.click();
+
+			const historyScopedInfoMessage = page.getByText("Chatverlauf pausiert");
+			await expect(historyScopedInfoMessage).toBeVisible();
+		},
+	);
+
+	testWithMockedLlm(
+		"Activating web search does not show chat history paused info message in a fresh chat",
+		async ({ page, isMobile }) => {
+			testWithMockedLlm.skip(isMobile === true, "desktop-only menu flow");
+
+			if (process.env.VITE_FEATURE_FLAG_WEB_SEARCH_ALLOWED !== "true") {
+				testWithMockedLlm.skip();
+			}
+
+			await page.goto("/");
+
+			const chatOptionsButton = page.getByRole("button", {
+				name: "Weitere Funktionen aktivieren",
+			});
+			await chatOptionsButton.click();
+
+			const webSearchOption = page.getByRole("menuitemcheckbox", {
+				name: "Websuche auswählen",
+			});
+			await webSearchOption.click();
+
+			const chatInput = page.getByPlaceholder("Das Web durchsuchen");
+			await expect(chatInput).toBeVisible();
+
+			const historyScopedInfoMessage = page.getByText("Chatverlauf pausiert");
+			await expect(historyScopedInfoMessage).not.toBeVisible();
+		},
+	);
+
+	testWithMockedLlm(
+		"Manually deactivating web search clears chat history paused info message",
+		async ({ page, isMobile }) => {
+			testWithMockedLlm.skip(isMobile === true, "desktop-only menu flow");
+
+			if (process.env.VITE_FEATURE_FLAG_WEB_SEARCH_ALLOWED !== "true") {
+				testWithMockedLlm.skip();
+			}
+
+			await page.goto("/");
+
+			const chatInput = page.getByPlaceholder("Stellen Sie eine Frage");
+			await chatInput.fill("Hallo, wie geht es dir?");
+			await sendAndWaitForLLMResponse(page);
+
+			const answer = page.getByTestId("assistant-message-markdown-container");
+			await expect(answer).not.toBeEmpty();
+
+			const chatOptionsButton = page.getByRole("button", {
+				name: "Weitere Funktionen aktivieren",
+			});
+			await chatOptionsButton.click();
+
+			const webSearchOption = page.getByRole("menuitemcheckbox", {
+				name: "Websuche auswählen",
+			});
+			await webSearchOption.click();
+
+			const historyScopedInfoMessage = page.getByText("Chatverlauf pausiert");
+			await expect(historyScopedInfoMessage).toBeVisible();
+
+			const webSearchPill = page.getByRole("button", {
+				name: "Websuche entfernen",
+			});
+			await webSearchPill.click();
+
+			await expect(historyScopedInfoMessage).not.toBeVisible();
+		},
+	);
+
+	testDesktopOnly(
+		"Adding document replaces chat history paused info message with auto-deactivation info message",
+		async ({ page }) => {
+			if (process.env.VITE_FEATURE_FLAG_WEB_SEARCH_ALLOWED !== "true") {
+				testDesktopOnly.skip();
+			}
+
+			await page.goto("/");
+
+			const chatInput = page.getByPlaceholder("Stellen Sie eine Frage");
+			await chatInput.fill("Hallo, wie geht es dir?");
+			await sendAndWaitForLLMResponse(page);
+
+			const answer = page.getByTestId("assistant-message-markdown-container");
+			await expect(answer).not.toBeEmpty();
+
+			const chatOptionsButton = page.getByRole("button", {
+				name: "Weitere Funktionen aktivieren",
+			});
+			await chatOptionsButton.click();
+
+			const webSearchOption = page.getByRole("menuitemcheckbox", {
+				name: "Websuche auswählen",
+			});
+			await webSearchOption.click();
+
+			const historyScopedInfoMessage = page.getByText("Chatverlauf pausiert");
+			await expect(historyScopedInfoMessage).toBeVisible();
+
+			const addDocumentToChatButton = page
+				.getByRole("listitem")
+				.filter({ hasText: defaultDocumentName })
+				.getByLabel("In den Chat");
+			await addDocumentToChatButton.click();
+
+			const toolDeactivatedInfoMessage = page.getByText(
+				"Websuche wurde automatisch deaktiviert.",
+			);
+			await expect(toolDeactivatedInfoMessage).toBeVisible();
+			await expect(historyScopedInfoMessage).not.toBeVisible();
+		},
+	);
+
 	testWithLoggedInUser(
 		"Should not be able to send another message with enter while waiting for a response",
 		async ({ page }) => {
